requirements.txt
discord.py
asyncpg
python-dotenv
colorlog
tortoise-orm